Move individuals table to separate function
